afternoon class

+++ c_example/genderRatio.c
@@ -10,9 +10,9 @@ int main(void){
     printf("woman count : ");
     scanf("%d", &woman);
 
-    sum = man + woman;
-    manRate = man/sum*100;
-    womanRate = woman/sum*100;
+    sum = (double)man + (double)woman;
+    manRate = (double)man/sum*100;
+    womanRate = (double)woman/sum*100;
 
     printf("man count %d, woman count %d \n", man, woman);
     printf("total count %.f \nmanRate %.2f%% \nwomanRate %.2f%% \n", sum, manRate, womanRate);
